+++ extension/src/cli/dvc/contract.ts
@@ -1,7 +1,7 @@
 import { Plot } from '../../plots/webview/contract'
 
 export const MIN_CLI_VERSION = '2.52.0'
-export const LATEST_TESTED_CLI_VERSION = '2.53.0'
+export const LATEST_TESTED_CLI_VERSION = '2.54.0'
 export const MAX_CLI_VERSION = '3'
 
 type ErrorContents = { type: string; msg: string }